# perf: do not use regex to escape expressions