See https://github.com/gurgunday/ghtml/pull/16#issuecomment-2156104795

closes https://github.com/gurgunday/ghtml/issues/15