@gurgunday 2024-06-07
Interesting, on macOS, the PR is faster by a big margin