@gurgunday 2024-06-08
Kitajs benchmark:

OLD:

```sh
benchmark        time (avg)             (min … max)       p75       p99      p999
--------------------------------------------------- -----------------------------
• Real World Scenario
--------------------------------------------------- -----------------------------
KitaJS/Html     544 µs/iter     (381 µs … 4'877 µs)    422 µs  1'885 µs  4'802 µs
Typed Html    1'950 µs/iter   (1'655 µs … 2'773 µs)  2'200 µs  2'633 µs  2'773 µs
VHtml         2'527 µs/iter   (2'297 µs … 3'608 µs)  2'768 µs  3'517 µs  3'608 µs
React JSX     6'580 µs/iter  (5'995 µs … 10'652 µs)  6'859 µs 10'652 µs 10'652 µs
Preact          966 µs/iter     (688 µs … 3'018 µs)    800 µs  2'451 µs  3'018 µs
React         6'598 µs/iter   (5'997 µs … 7'851 µs)  7'210 µs  7'851 µs  7'851 µs
Common Tags   3'297 µs/iter   (2'915 µs … 5'016 µs)  3'190 µs  4'954 µs  5'016 µs
Ghtml           387 µs/iter     (308 µs … 2'539 µs)    348 µs  2'295 µs  2'503 µs
JSXTE         5'169 µs/iter   (4'961 µs … 9'826 µs)  5'191 µs  5'999 µs  9'826 µs

summary for Real World Scenario
  Ghtml
   1.41x faster than KitaJS/Html
   2.5x faster than Preact
   5.04x faster than Typed Html
   6.52x faster than VHtml
   8.51x faster than Common Tags
   13.35x faster than JSXTE
   16.99x faster than React JSX
   17.04x faster than React

• Component Creation
--------------------------------------------------- -----------------------------
KitaJS/Html     651 µs/iter     (505 µs … 2'261 µs)    529 µs  1'781 µs  2'261 µs
Typed Html    1'017 µs/iter     (823 µs … 2'580 µs)    875 µs  2'250 µs  2'580 µs
VHtml         1'815 µs/iter   (1'536 µs … 3'295 µs)  1'855 µs  3'093 µs  3'295 µs
React JSX     9'184 µs/iter  (8'190 µs … 14'233 µs)  9'549 µs 14'233 µs 14'233 µs
Preact          818 µs/iter     (654 µs … 3'283 µs)    769 µs  2'135 µs  3'283 µs
React         8'997 µs/iter   (8'179 µs … 9'858 µs)  9'424 µs  9'858 µs  9'858 µs
Common Tags   1'845 µs/iter   (1'723 µs … 2'436 µs)  2'021 µs  2'334 µs  2'436 µs
Ghtml           646 µs/iter     (628 µs … 1'098 µs)    633 µs  1'044 µs  1'098 µs
JSXTE         4'733 µs/iter  (4'351 µs … 10'337 µs)  4'749 µs  6'071 µs 10'337 µs

summary for Component Creation
  Ghtml
   1.01x faster than KitaJS/Html
   1.27x faster than Preact
   1.57x faster than Typed Html
   2.81x faster than VHtml
   2.86x faster than Common Tags
   7.32x faster than JSXTE
   13.92x faster than React
   14.21x faster than React JSX

• Attributes Serialization
--------------------------------------------------- -----------------------------
KitaJS/Html  20'114 ns/iter    (18'833 ns … 508 µs) 19'417 ns 24'208 ns    257 µs
Typed Html   79'511 ns/iter    (74'416 ns … 498 µs) 76'041 ns    298 µs    380 µs
VHtml        91'537 ns/iter    (87'000 ns … 528 µs) 88'541 ns    148 µs    381 µs
React JSX    77'500 ns/iter    (70'708 ns … 567 µs) 74'750 ns    192 µs    423 µs
Preact       19'032 ns/iter    (17'458 ns … 515 µs) 18'167 ns 35'208 ns    274 µs
React        77'153 ns/iter    (70'833 ns … 528 µs) 74'125 ns    198 µs    421 µs
Common Tags  52'154 ns/iter    (48'166 ns … 593 µs) 50'541 ns    100 µs    362 µs
Ghtml        46'792 ns/iter    (44'083 ns … 490 µs) 46'125 ns 89'333 ns    329 µs
JSXTE        29'756 ns/iter    (25'792 ns … 517 µs) 27'375 ns 82'833 ns    400 µs

summary for Attributes Serialization
  Preact
   1.06x faster than KitaJS/Html
   1.56x faster than JSXTE
   2.46x faster than Ghtml
   2.74x faster than Common Tags
   4.05x faster than React
   4.07x faster than React JSX
   4.18x faster than Typed Html
   4.81x faster than VHtml
```

NEW:

```sh
benchmark        time (avg)             (min … max)       p75       p99      p999
--------------------------------------------------- -----------------------------
• Real World Scenario
--------------------------------------------------- -----------------------------
KitaJS/Html     507 µs/iter     (389 µs … 1'836 µs)    412 µs  1'221 µs  1'557 µs
Typed Html    1'901 µs/iter   (1'649 µs … 2'670 µs)  2'138 µs  2'600 µs  2'670 µs
VHtml         2'480 µs/iter   (2'301 µs … 3'487 µs)  2'732 µs  3'236 µs  3'487 µs
React JSX     6'459 µs/iter   (6'014 µs … 7'654 µs)  6'881 µs  7'654 µs  7'654 µs
Preact          972 µs/iter     (689 µs … 2'777 µs)    817 µs  2'479 µs  2'777 µs
React         6'597 µs/iter   (6'049 µs … 8'109 µs)  6'885 µs  8'109 µs  8'109 µs
Common Tags   3'078 µs/iter   (2'870 µs … 4'420 µs)  3'051 µs  4'276 µs  4'420 µs
Ghtml           221 µs/iter     (183 µs … 1'454 µs)    203 µs  1'184 µs  1'390 µs
JSXTE         4'404 µs/iter   (3'738 µs … 6'206 µs)  4'458 µs  5'468 µs  6'206 µs

summary for Real World Scenario
  Ghtml
   2.29x faster than KitaJS/Html
   4.39x faster than Preact
   8.59x faster than Typed Html
   11.2x faster than VHtml
   13.91x faster than Common Tags
   19.9x faster than JSXTE
   29.18x faster than React JSX
   29.81x faster than React

• Component Creation
--------------------------------------------------- -----------------------------
KitaJS/Html     563 µs/iter     (505 µs … 1'178 µs)    521 µs    977 µs  1'107 µs
Typed Html      907 µs/iter     (824 µs … 1'535 µs)    853 µs  1'370 µs  1'535 µs
VHtml         1'645 µs/iter   (1'533 µs … 2'875 µs)  1'687 µs  2'500 µs  2'875 µs
React JSX     9'016 µs/iter  (8'177 µs … 10'329 µs)  9'496 µs 10'329 µs 10'329 µs
Preact          770 µs/iter     (656 µs … 1'791 µs)    706 µs  1'550 µs  1'791 µs
React         9'140 µs/iter  (8'233 µs … 12'365 µs)  9'555 µs 12'365 µs 12'365 µs
Common Tags   1'841 µs/iter   (1'724 µs … 2'658 µs)  1'988 µs  2'364 µs  2'658 µs
Ghtml           447 µs/iter       (430 µs … 932 µs)    435 µs    824 µs    917 µs
JSXTE         4'470 µs/iter   (4'274 µs … 5'194 µs)  4'511 µs  4'945 µs  5'194 µs

summary for Component Creation
  Ghtml
   1.26x faster than KitaJS/Html
   1.72x faster than Preact
   2.03x faster than Typed Html
   3.68x faster than VHtml
   4.12x faster than Common Tags
   10x faster than JSXTE
   20.17x faster than React JSX
   20.45x faster than React

• Attributes Serialization
--------------------------------------------------- -----------------------------
KitaJS/Html  20'126 ns/iter    (18'708 ns … 496 µs) 19'375 ns 26'083 ns    316 µs
Typed Html   79'555 ns/iter    (74'250 ns … 429 µs) 75'875 ns    312 µs    399 µs
VHtml        92'170 ns/iter    (87'083 ns … 459 µs) 88'708 ns    179 µs    410 µs
React JSX    78'670 ns/iter    (69'750 ns … 827 µs) 74'750 ns    277 µs    510 µs
Preact       19'414 ns/iter    (17'458 ns … 468 µs) 18'375 ns 46'667 ns    294 µs
React        75'486 ns/iter    (69'875 ns … 430 µs) 73'291 ns    158 µs    358 µs
Common Tags  52'608 ns/iter    (48'458 ns … 508 µs) 50'708 ns    121 µs    344 µs
Ghtml        29'061 ns/iter    (27'250 ns … 398 µs) 28'500 ns 49'250 ns    292 µs
JSXTE        29'119 ns/iter    (25'833 ns … 415 µs) 27'209 ns 54'875 ns    352 µs

summary for Attributes Serialization
  Preact
   1.04x faster than KitaJS/Html
   1.5x faster than Ghtml
   1.5x faster than JSXTE
   2.71x faster than Common Tags
   3.89x faster than React
   4.05x faster than React JSX
   4.1x faster than Typed Html
   4.75x faster than VHtml
```