@gurgunday 2024-06-08
@gunjam what do you think?

Kita benchmarks show significant improvement!

Do you have any suggestions to further improve the results?