@gunjam 2024-06-09
Hey! looks great, nice work!

Also seeing good improvements on my first pass component code with your changes:

main
```
benchmark                         time (avg)             (min … max)       p75       p99      p999
-------------------------------------------------------------------- -----------------------------
govukButton - start now link     306 ns/iter       (268 ns … 624 ns)    313 ns    372 ns    624 ns
govukButton - input              333 ns/iter       (297 ns … 423 ns)    344 ns    390 ns    423 ns
govukButton - button             358 ns/iter       (338 ns … 438 ns)    366 ns    396 ns    438 ns
```
this pr
```
benchmark                         time (avg)             (min … max)       p75       p99      p999
-------------------------------------------------------------------- -----------------------------
govukButton - start now link     194 ns/iter       (171 ns … 364 ns)    198 ns    235 ns    339 ns
govukButton - input              212 ns/iter       (193 ns … 295 ns)    221 ns    252 ns    291 ns
govukButton - button             268 ns/iter       (240 ns … 346 ns)    279 ns    312 ns    345 ns
```

I can't see how you could make it any better, thanks again for your hard work